* DbContext: - refactor changes tracking, - update, insert, delete use bulk when possible - refactor soft delete - refactor tvp

+++ packages/elcy-core/src/Queryable/QueryExpression/EntityExpression.ts
@@ -75,7 +75,7 @@ export class EntityExpression<TE extends object = object> implements IEntityExpr
     }
     public get versionColumn() {
         if (typeof this._versionColumn === "undefined") {
-            this._versionColumn = !this.metaData || !this.metaData.versionColumn ? null : this.columns.find((o) => o.propertyName === this.metaData.versionColumn.propertyName) as IColumnExpression<TE, Uint8Array>;
+            this._versionColumn = !this.metaData || !this.metaData.versionColumn ? null : this.columns.find((o) => o.propertyName === this.metaData.versionColumn.propertyName) as IColumnExpression<TE, bigint>;
         }
         return this._versionColumn;
     }
@@ -99,7 +99,7 @@ export class EntityExpression<TE extends object = object> implements IEntityExpr
     private _metaData: IEntityMetaData<TE>;
     private _modifiedColumn: IColumnExpression<TE, Date>;
     private _primaryColumns: IColumnExpression<TE>[];
-    private _versionColumn: IColumnExpression<TE, Uint8Array>;
+    private _versionColumn: IColumnExpression<TE, bigint>;
     public clone(replaceMap?: Map<IExpression, IExpression>): EntityExpression<TE> {
         if (!replaceMap) {
             replaceMap = new Map();
